% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bond_cf}
\alias{bond_cf}
\title{Generate bond's cash flows}
\usage{
bond_cf(value_date, mty_date, redem_value, cpn_rate, cpn_freq)
}
\arguments{
\item{value_date}{the value and maturity date of the bond}

\item{mty_date}{the value and maturity date of the bond}

\item{redem_value}{the redemption value, coupon rate and coupon frequency of the bond.
Note that the \strong{frequency} can only be one of 1, 2, 4, 0 (pay at mature)}

\item{cpn_rate}{the redemption value, coupon rate and coupon frequency of the bond.
Note that the \strong{frequency} can only be one of 1, 2, 4, 0 (pay at mature)}

\item{cpn_freq}{the redemption value, coupon rate and coupon frequency of the bond.
Note that the \strong{frequency} can only be one of 1, 2, 4, 0 (pay at mature)}
}
\description{
Generate bond's cash flows
}
