% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bond_result}
\alias{bond_result}
\title{Calculate the Bond's YTM, Maclay Duration, Modified Duration}
\usage{
bond_result(
  value_date,
  mty_date,
  redem_value,
  cpn_rate,
  cpn_freq,
  ref_date,
  clean_price
)
}
\arguments{
\item{value_date, mty_date}{the value and maturity date of the bond}

\item{redem_value, cpn_rate, cpn_freq}{the redemption value, coupon rate and coupon frequency of the bond.
Note that the \strong{frequency} can only be one of 1, 2, 4, 0 (pay at mature)}

\item{ref_date, clean_price}{the reference date and the clean price that used to calculate the bond results}
}
\value{
a double vector with 3 elements: ytm, macd and modd
}
\description{
Calculate the Bond's YTM, Maclay Duration, Modified Duration
}
