% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtn.R
\name{create_rtn}
\alias{create_rtn}
\title{Create Rtn Object}
\usage{
create_rtn(ids, dates, mvs, pls)
}
\arguments{
\item{ids}{an integer vector, the ID each row belongs to}

\item{dates}{a Date vector, the}
}
\description{
By providing a "group" (\code{ids}) of \code{dates}, \code{mvs} and \code{pls},
calucating the time-weighted rate of rtn (twrr) or modified
dietz rate of return (dietz).
}
\details{
All the input vector must be the same length.
}
