% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-bond.R
\name{fixed_bond}
\alias{fixed_bond}
\title{Create Fixed Bond Object}
\usage{
fixed_bond(value_date, mty_date, redem_value, cpn_rate, cpn_freq)
}
\arguments{
\item{value_date, mty_date}{the value and maturity date of the bond}

\item{redem_value, cpn_rate, cpn_freq}{the redemption value, coupon rate and coupon frequency of the bond.
Note that the \strong{frequency} can only be one of 1, 2, 4, 0 (pay at mature)}

\item{ref_date, clean_price}{the reference date and the clean price that used to calculate the bond results}
}
\description{
Create Fixed Bond Object
}
\details{
a double vector with 3 elements: ytm, macd and modd
}
