% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtn.R
\name{rtn}
\alias{rtn}
\title{Create Rtn Object}
\usage{
rtn(id, date, mv, pl)
}
\arguments{
\item{id}{an integer vector, the ID of each row belongs to}

\item{date}{a Date vector, the reference date of each row}

\item{mv, pl}{a Double vector, the market value and the PnL of each day}
}
\description{
By providing a "group" (\code{ids}) of \code{dates}, \code{mvs} and \code{pls},
calucating the time-weighted rate of rtn (twrr) or modified
dietz rate of return (dietz).
}
\details{
All the input vector must be 1 or the same length.
}
