% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-bond.R
\name{fixed_bond}
\alias{fixed_bond}
\title{Create Fixed Bond Object}
\usage{
fixed_bond(value_date, mty_date, redem_value, cpn_rate, cpn_freq)
}
\arguments{
\item{value_date, mty_date}{the value and maturity date of the bond}

\item{redem_value, cpn_rate, cpn_freq}{the redemption value, coupon rate and coupon frequency of the bond.
Note that the \strong{frequency} can only be one of 1, 2, 4, 0 (pay at mature)}
}
\description{
Create Fixed Bond Object
}
\details{
it returns an environment containing the following objects:
\itemize{
\item \code{.self}: an external pointer of the Rust object.
\item \code{len()}: a function returns the length of the internal bonds object.
\item \code{ytm_dur(ref_date, clean_price)}: a function returns a data.frame, with three columns,
YTM, MODD and MACD.
\item \code{cf(ref_date)}: a function returns the schedualed bond cashflows, in \code{xts} format.
}
}
\note{
\itemize{
\item all arguments must be the same length or 1.
\item The date input will be converted to Date object via \code{\link[ymd:ymd]{ymd::ymd()}}.
}
}
